feat: extracted ui logic into self contained package
