+++ packages/cli/src/commands/list/listFactory.ts
@@ -1,26 +1,17 @@
+import { ExitCodes, Status } from '@bundler/common';
+import { Renderer, createTerminalStreamer, ListStyleRequestBuilder as Builder } from '@bundler/terminal-ui';
 import { FactoryFunction } from 'tsyringe';
 import { Logger } from '@map-colonies/js-logger';
 import { GithubRepository, RepositoryType, IGithubClient } from '@bundler/github';
 import { GITHUB_ORG } from '@bundler/core';
 import { Arguments, Argv, CommandModule } from 'yargs';
-import chalk from 'chalk';
 import { GlobalArguments } from '../../cliBuilderFactory';
-import { ExitCodes, EXIT_CODE, LifeCycle, SERVICES, Status } from '../../common/constants';
+import { EXIT_CODE, SERVICES, TERMINAL_STREAM } from '../../common/constants';
 import { checkWrapper } from '../../wrappers/check';
-import { oldCreateTerminalStreamer } from '../../ui/terminalStreamer';
-import { ExtendedColumnifyOptions, style } from '../../ui/styler';
-import { command, describe, PREFIX } from './constants';
+import { command, describe } from './constants';
 import { visibilityTokenImplicationCheck } from './checks';
 
-const columnifyOptions: ExtendedColumnifyOptions = {
-  align: 'left',
-  preserveNewLines: true,
-  columns: ['name', 'language', 'topics'],
-  showHeaders: true,
-  alternateChalks: [chalk.bold.hex('#a5d4d3'), chalk.cyanBright],
-};
-
-interface Listed extends Pick<GithubRepository, 'name' | 'language' | 'topics'> {
+export interface Listed extends Pick<GithubRepository, 'name' | 'language' | 'topics'> {
   status: Status;
 }
 
@@ -58,34 +49,25 @@ export const listCommandFactory: FactoryFunction<CommandModule<GlobalArguments, 
     logger.debug({ msg: 'executing command', command, args: { visibility, topics }, filter });
 
     try {
+      const renderer = new Renderer(createTerminalStreamer(TERMINAL_STREAM));
+      const builder = new Builder();
+      renderer.current = builder.build({ list: [], status: Status.PENDING });
+
       const filtered: Listed[] = [];
 
-      let cycle = LifeCycle.PRE;
-
-      const getData = (): string => {
-        const status = cycle === LifeCycle.POST ? Status.SUCCESS : Status.PENDING;
-        const prefix = { content: PREFIX(command), isBold: true, status };
-        const spinner = { level: 3, status };
-
-        if (status === Status.PENDING && filtered.length === 0) {
-          return style({ prefix, main: [spinner] });
+      for await (const repos of githubClient.listRepositoriesGenerator(GITHUB_ORG, visibility, filter)) {
+        if (repos.length === 0) {
+          const request = builder.build({ list: filtered, status: Status.SUCCESS });
+          renderer.current = request;
+          break;
         }
 
-        const columns = { level: 3, status, content: { config: columnifyOptions, data: filtered } };
-        const main = cycle === LifeCycle.POST ? [columns] : [columns, spinner];
-        const suffix = { level: 3, status, isBold: true, content: ` listed ${filtered.length} repositories ` };
-        return style({ prefix: { content: PREFIX(command), isBold: true, status }, main, suffix });
-      };
-
-      oldCreateTerminalStreamer(process.stderr, getData);
-
-      for await (const repos of githubClient.listRepositoriesGenerator(GITHUB_ORG, visibility, filter)) {
-        filtered.push(...repos.map((r) => ({ name: r.name, language: r.language, topics: r.topics, status: Status.SUCCESS })));
+        filtered.push(...repos.map((r) => ({ name: r.name, language: r.language, topics: r.topics, status: Status.PENDING })));
+        const request = builder.build({ list: filtered, status: Status.PENDING });
+        renderer.current = request;
       }
 
       // logger.debug({ msg: 'got repositories', count: repos.length });
-
-      cycle = LifeCycle.POST;
 
       dependencyContainer.register(EXIT_CODE, { useValue: ExitCodes.SUCCESS });
     } catch (error) {
