+++ packages/common/README.md
@@ -1,0 +1,1 @@
+# terminal-ui
