+++ packages/common/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@bundler/cli",
+  "name": "@bundler/common",
   "version": "1.0.0",
-  "description": "bundler cli",
+  "description": "bundler common",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "scripts": {
@@ -16,45 +16,31 @@
     "release": "standard-version",
     "test": "npm run test:unit && npm run test:integration",
     "prebuild": "npm run clean",
-    "build": "tsc --project tsconfig.build.json && npm run assets:copy",
+    "build": "tsc --project tsconfig.build.json",
     "start": "npm run build && cd dist && node ./index.js",
-    "assets:copy": "copyfiles ./package.json dist",
-    "clean": "rimraf dist"
+    "clean": "rimraf dist",
+    "prepack": "npm run build"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/MapColonies/bundler.git"
+    "url": ""
   },
   "author": "MapColonies",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/MapColonies/bundler/issues"
+  "files": [
+    "dist/**/*"
+  ],
+  "directories": {
+    "test": "tests"
   },
-  "homepage": "https://github.com/MapColonies/bundler#readme",
+  "homepage": "",
   "dependencies": {
-    "@bundler/common": "*",
-    "@bundler/core": "*",
-    "@bundler/github": "*",
-    "@bundler/terminal-ui": "*",
-    "@map-colonies/js-logger": "^0.0.5",
-    "@map-colonies/read-pkg": "0.0.1",
-    "chalk": "^4.1.2",
-    "cli-spinners": "^2.7.0",
-    "columnify": "^1.6.0",
-    "config": "^3.3.8",
-    "reflect-metadata": "^0.1.13",
-    "tsyringe": "^4.7.0",
-    "yargs": "^17.6.2"
   },
   "devDependencies": {
     "@map-colonies/eslint-config": "^3.1.3",
     "@map-colonies/prettier-config": "0.0.1",
-    "@types/columnify": "^1.5.1",
-    "@types/config": "^3.3.0",
     "@types/jest": "^26.0.19",
-    "@types/node": "^18.11.18",
-    "@types/yargs": "^17.0.18",
-    "copyfiles": "^2.4.1",
+    "@types/node": "^14.14.12",
     "eslint": "^8.30.0",
     "jest": "^26.6.3",
     "jest-create-mock-instance": "^1.1.0",
