+++ packages/github/src/client/index.ts
@@ -79,9 +79,6 @@ export class GithubClient implements IGithubClient {
   }
 
   public async ping(): Promise<void> {
-    // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-    await new Promise((resolve) => setTimeout(resolve, 5000));
-
     await this._octokit.rest.orgs.get({
       org: GITHUB_ORG,
     });
