+++ packages/terminal-ui/.prettierignore
@@ -1,0 +1,13 @@
+dist
+.github
+.gitignore
+CHANGELOG.md
+commitlint.config.js
+package-lock.json
+package.json
+README.md
+tsbuildconfig.json
+tsconfig.json
+coverage
+reports
+docs
