+++ packages/terminal-ui/src/style/requestBuilders/index.ts
@@ -1,0 +1,7 @@
+import { StyleRequest } from '../styler';
+
+export const PREFIX = (command: string): string => ` > executing ${command} command `;
+
+export abstract class StyleRequestBuilder {
+  public abstract build(data: unknown): StyleRequest;
+}
