+++ packages/terminal-ui/tsconfig.lint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../tsconfig.lint.json"
+}
